hle: kernel: Separate KScheduler from GlobalSchedulerContext class.
