+++ src/tests/CMakeLists.txt
@@ -2,7 +2,6 @@ add_executable(tests
     common/bit_field.cpp
     common/bit_utils.cpp
     common/fibers.cpp
-    common/multi_level_queue.cpp
     common/param_package.cpp
     common/ring_buffer.cpp
     core/arm/arm_test_common.cpp
